bot/pyproject.toml
[tool.poetry]
name = "mykolaiv_city_bot_client"
version = "0.1.0"
description = ""
authors = ["cheerfulleg <yolginivan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
aiogram = "^2.20"
httpx = "^0.22.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

